Added WIP server start code Renamed SteamClient to SteamManager
