+++ UnityProject/Assets/Plugins/Facepunch.Steamworks/Utility/SteamManager.cs.meta
